Larger integers.

+++ Source/biblioteq_photographcollection.h
@@ -70,10 +70,10 @@ class biblioteq_photographcollection: public QMainWindow, public biblioteq_item
 			      QGraphicsPixmapItem *item,
 			      const int percent);
   void loadPhotographFromItemInNewWindow(QGraphicsPixmapItem *item);
-  void setSceneRect(const int size);
+  void setSceneRect(const qint64 size);
   void showPhotographs(const int page);
   void storeData(void);
-  void updateTablePhotographCount(const int count);
+  void updateTablePhotographCount(const qint64 count);
 
  private slots:
   void slotAddItem(void);
